+++ changelog.md
@@ -10,6 +10,7 @@
 
 - Updated to **.NET 10**
   - Increased overall performance
+- #1469 Show Password Policy and Live Validation on Customer Registration.
 - **Wallet**
   - Wallet credits can be added to the shopping cart anywhere using a PageBuilder block.
   - An option has been added to display wallet credit amounts as buttons when purchasing credit.